+++ files/private-chef-cookbooks/private-chef/recipes/keepalived.rb
@@ -19,9 +19,10 @@ keepalived_log_dir = node['private_chef']['keepalived']['log_directory']
 end
 
 directory keepalived_log_dir do
-  owner node['private_chef']['user']['username']
+  owner OmnibusHelper.new(node).ownership['owner']
+  group OmnibusHelper.new(node).ownership['group']
   recursive true
-  mode "0700"
+  mode node['private_chef']['service_dir_perms']
 end
 
 template File.join(keepalived_etc_dir, "keepalived.conf") do