+++ files/private-chef-cookbooks/private-chef/recipes/nginx.rb
@@ -31,8 +31,9 @@ nginx_addon_dir = File.join(nginx_etc_dir, "addon.d")
   nginx_tempfile_dir
 ].each do |dir_name|
   directory dir_name do
-    owner node['private_chef']['user']['username']
-    mode '0700'
+    owner OmnibusHelper.new(node).ownership['owner']
+    group OmnibusHelper.new(node).ownership['group']
+    mode node['private_chef']['service_dir_perms']
     recursive true
   end
 end
@@ -93,8 +94,9 @@ remote_directory nginx_html_dir do
   files_owner "root"
   files_group "root"
   files_mode "0644"
-  owner node['private_chef']['user']['username']
-  mode "0700"
+  owner OmnibusHelper.new(node).ownership['owner']
+  group OmnibusHelper.new(node).ownership['group']
+  mode node['private_chef']['service_dir_perms']
 end
 
 nginx_config = File.join(nginx_etc_dir, "nginx.conf")